% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter2}
\alias{filter2}
\title{Filter rows}
\usage{
filter2(df, ind)
}
\arguments{
\item{df}{A data frame.}

\item{ind}{A vector of indices, or a logical vector.}
}
\value{
A data frame with the subsetted rows.
}
\description{
Filter rows of a data frame
}
\examples{
filter2(iris, 1:3)
filter2(iris, Species == "setosa")
}
