% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr3-package.R
\docType{package}
\name{minidplyr3-package}
\alias{minidplyr3}
\alias{minidplyr3-package}
\title{minidplyr3: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/privefl/minidplyr3}
  \item Report bugs at \url{https://github.com/privefl/minidplyr3/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
