% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Subset columns}
\usage{
select(df, col)
}
\arguments{
\item{df}{A data frame.}

\item{col}{Either a character vector with the names of columns to keep
or an integer vector with the positions of the columns to keep.}
}
\value{
A data frame with the columns kept.
}
\description{
Subset columns
}
\examples{
select(iris, 1:4)
select(iris, c("Sepal.Length", "Sepal.Width"))
}
