% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr3-package.R
\docType{package}
\name{minidplyr3-package}
\alias{minidplyr3}
\alias{minidplyr3-package}
\title{minidplyr3: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
